fix: Allow paging through existing results after changing search field
With the previous implementation, paging through results used the search term in the input field to fetch the next or previous page of results even if the user hadn't clicked "Search". With that implementation, a user could search for something, be shown results, change the value in the search input field without clicking "Search", click the next or previous page button, and they would be shown results for the updated value in the search input field even though they didn't actually click the "Search" button. This led to surprising behavior and, in some cases, errors.

This commit creates a separate piece of state to track the last search term that the user actually searched with. Paging uses that search term instead of the current value in the search input field. This allows a user to change the value in the search input field and it won't affect displayed results until they actually click "Search".