feature: Render edit modal alongside users in search results
This implementation renders a separate edit modal for each of the users displayed in search results.

I considered rendering a single edit modal to be used for all of the displayed users; that seemed to be cleaner as far as the DOM is concerned. Rendering a modal for each user seemed to simplify the code though, and that seems more valuable. Also, rendering a separate modal for each entity seems to be the pattern we use elsewhere in our applications.