feature: Include emails when fetching account users
The `include[]` param is not documented in the [Canvas API docs](https://canvas.instructure.com/doc/api/users.html#method.users.api_index) but it is supported. Since it's supported by other endpoints, we concluded it's probably reasonable to expect it to continue to be there, but it still might be a risk since it's undocumented. Or their docs might just be incomplete.