fix: Fix bug in SearchPage
We were checking the search term length and verifying that it wasn't too short before submitting the search, but we were updating the `resultsSearchTerm` state property whether or not it was long enough and we had performed the search.

This meant that, if a user tried to search with a search term that was too short, they would be correctly prevented from performing that search. However, if they then tried to page through their current results, the callback function on the paging buttons would use the search term that was too short (because we updated state when we shouldn't have) and we would get an error response from the server.