feature: Only allow admins to modify users from their accounts
Before modifying user data in the `update` action, we verify that the user belongs to the account (or one of the account's sub-accounts) by searching for the user using the `accounts/:canvas_account_id/users` Canvas API endpoint.

If this request to the Canvas API comes back empty, we can assume that the user doesn't belong to the account the admin is authorized to access and we should reject the request to our application.