feature: Don't nest `canvas_users` endpoint under `canvas_accounts`
Originally, we nested the `canvas_users` endpoint under `canvas_accounts`. e.g. `/canvas_accounts/123/canvas_users`. This made sense because actions were performed on users (or a single user) within an account. That's still how actions are performed, but we're no longer allowing the client to specify which account they want operate within because it creates a security risk. Users (who should be admin users) are only allowed to list and modify users which belong to the account (or its child-accounts) that they administer.

To ensure admins are only accessing users they have permissions to access, we're using the Canvas account ID we get in the LTI launch instead. Because we're using the account ID from the LTI launch, it no longer makes sense to have clients specify the Canvas account ID in the URL.

We're also renaming the endpoint to `canvas_account_users` to make it clearer that the endpoint does not operate on all Canvas users but just users of a certain Canvas account.