feature: Add better handling for Canvas API errors
Because we're making multiple calls to the Canvas API to update all of the different user attributes, it's possible that we might successfully update the user's name and/or email and then fail to update the user's login ID, password and/or SIS ID. That's why we have some additional logic and custom messaging as contained in the `render_update_user_exception` method. We want to make sure the user understands exactly what information we failed to update.