feature: Change `create_by_diffing_attrs` to a bang method
Instead of calling `ActiveRecord.create`, we'll call `ActiveRecord.create!`. I think this makes sense because we're using this method to log changes to sensitive user data. If something goes wrong and the logging record isn't created, we want to know about it.

If a non-bang version of the method is desired at some point in the future, it would be very easy to create one.