refactor: Ensure logging user changes always happens
With the previous implementation, we weren't logging the user changes if an unhandled exception occurred. We want to ensure that the logging record is created no matter how the action may fail.

To be clear, if a before action of some kind fails, we won't create a logging record. I think this is reasonable. If, say, someone tries to hit the update action but is unauthorized or has an invalid token, I don't think we need to try and create a logging record in that type of scenario. If someone's request successfully makes it to the update action though, we want to create a logging record in all scenarios after that.