feature: Clear success messages after 4 seconds
The code in the `SuccessMessages` component that clears the messages is pretty simple. It just sets up a 4 second timeout, and it doesn't check if there is an existing timeout. This means, that if an additional success message is added and rendered before the original timeout expires, that additional message won't be displayed for the full 4 seconds; it will only be displayed until the original timeout expires and all the messages are cleared.

I think this is an acceptable implementation though. At least for now, we're only using this component to render success messages from updating a user. It would be very challenging for the admin to update multiple users within 4 seconds. The update request itself usually takes around 2 seconds. I can't see a possible scenario where a user would manage to pull this off if they're using the application like it's intended to. I'm not saying an admin couldn't do it; they could if they really wanted to. I'm just saying that I don't see value in supporting this scenario.