feature: Add canvas user show action
This action will be used to determine whether or not a given user is an account admin. We may also use it to fetch additional user information not returned by the index action.

I considered adding a `is_user_account_admin` action or something like that to the controller. Adding a `show` action seems more predictable and user friendly to me though. It's certainly more RESTful.