feature: Display admin error in edit user modal
Once the edit user modal mounts, we check if the user prop has an is_account_admin attribute. If it doesn't, we make an API call to the back-end to get it. Once we have that attribute, if a user is an account admin, we disable the edit form and display an error message.

This will prevent people from wasting their time trying to update an account admin. Right now, they're informed that the user is an account admin after they submit the form. This implementation will let them know, hopefully, before they even begin editing the form.

If a full 10 search results are returned, we end up making 10 API calls to the back-end. This isn't awesome, and it's not very fast, but I think it's acceptable because it's all happening in the background.