feature: Announce confirmation message with aria-live
The [MDN docs](https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/ARIA_Live_Regions) state, "Assistive technologies will announce dynamic changes in the content of a live region. The live region must first be present (and usually empty), so that the browser and assistive technologies are aware of it. Any subsequent changes are then announced."

That's why I moved the paragraph out of the conditional so it's always present in the DOM even if it's empty.