feature: Improve Canvas user change logging
This updates the `CanvasUserChange` model to only include an attribute in the `failed_attributes` list if the new value for that attribute was different from the original value. This will improve consistency between the `failed_attributes` list and the individual attribute fields; the individual attribute fields are only populated if the value is changed.

With the previous implementation, you could have `NULL` in an individual attribute field but then see that attribute in the `failed_attributes` list. This would happen in a scenario where the API request to Canvas to update that certain attribute failed, but the attribute wasn't changed. So the attribute was included in the `failed_attributes` list but the individual attribute field was left blank because the value didn't change. I find this potentially confusing. Now, you will only see an attribute in the `failed_attributes` list if it's corresponding individual attribute field is also populated. That individual field, of course, would have `success: false`.

Hopefully, these changes will make the `CanvasUserChange` records more understandable.