feature: Avoid unnecessary Canvas API calls when updating user
When updating a user, we make multiple API requests to Canvas to update the various pieces of user data that the tool allows you to update. However, if, for example, you only want to update the name of a user, there are a couple of API requests that are unnecessary because they're for updating other user data.

This commit changes the update action logic to only make those API calls when necessary. These changes have the potential to eliminate up to 2 of the 5 Canvas API calls made in the update action. (Technically, it could eliminate 3 calls, but that would only happen if none of the user data was changed. I don't see why that would happen during normal usage of the application.)