refactor: Don't search for Canvas users with their ID
The Canvas API docs for the [List users in account](https://canvas.instructure.com/doc/api/users.html#method.users.api_index) endpoint for the `search_term` parameter state, "Note that the API will prefer matching on canonical user ID if the ID has a numeric form. It will only search against other fields if non-numeric in form, or if the numeric value doesn't yield any matches." But that doesn't appear to be the case in my testing. I added the ID of one user to the name of another user and it pulled up both users when I searched using that ID.

There are a couple of places in the `CanvasAccountUsersController` where we search for users by ID when we want a specific user. Searching for users with their ID will cause issues if there's ever a user who has the ID of another user in their SIS ID or something like that.

Searching for users by ID also doesn't work if the user's ID is less than 3 characters. The API docs say the `search_term` must be at least 3 characters. I assumed this wouldn't apply to IDs, but I was incorrect. Searching for a user whose ID is 99 or less will give you a `400 Bad Request`.

Another problem with the current implementation is that, in the case where a user's ID exists in another user's SIS ID, email or some other attribute, we might erroneously conclude that a user is in the account and can thus be updated when they really aren't.

This commit removes the instances where we search for a user with their ID and instead makes use of the `/users/:id` Canvas API endpoint. Also, it searches using the user's login ID when we need to search for a specific user.