refactor: Combine before actions
We were fetching the Canvas user and validating that user was in the account in separate before actions, but I think it makes more sense to combine them. I don't foresee a scenario where we would want to fetch the Canvas user without also verifying that they're in the appropriate Canvas account, so combining the 2 methods into one simplifies the code a little bit and ensures we don't fetch the Canvas user without also verifying them.