fix: URI encode the search term
We observed a bug where search terms with a "+" character in them
weren't returning any results from the Canvas API even though there were
users that matched the search term.

The problem was that the plus sign was being sent to our back-end as
a space character instead of a literal plus sign. This search term with a
space instead of a plus was then sent to the Canvas API and there were
no matching results.

By URI encoding the search term before sending it to our back-end, we
get the results we expect.

I also tested searching with a number of other special characters
(~, !, /, #, ?, etc.) and they all worked fine.